More stars, Alt/Az in object details, filter only visible, double algebra

+++ app/src/main/java/io/github/marcocipriani01/telescopetouch/astronomy/OrbitalElements.java
@@ -21,7 +21,7 @@ import android.util.Log;
 import androidx.annotation.NonNull;
 
 import io.github.marcocipriani01.telescopetouch.TelescopeTouchApp;
-import io.github.marcocipriani01.telescopetouch.util.MathsUtils;
+import io.github.marcocipriani01.telescopetouch.maths.MathsUtils;
 
 import static java.lang.Math.abs;
 
@@ -40,20 +40,19 @@ import static java.lang.Math.abs;
  * @author Kevin Serafini
  * @author Brent Bryan
  */
-
 public class OrbitalElements {
 
     // calculation error
-    private final static float EPSILON = 1.0e-6f;
+    private final static double EPSILON = 1.0e-6f;
     private static final String TAG = TelescopeTouchApp.getTag(OrbitalElements.class);
-    public final float distance;       // Mean distance (AU)
-    public final float eccentricity;   // Eccentricity of orbit
-    public final float inclination;    // Inclination of orbit (AngleUtils.RADIANS)
-    public final float ascendingNode;  // Longitude of ascending node (AngleUtils.RADIANS)
-    public final float perihelion;     // Longitude of perihelion (AngleUtils.RADIANS)
-    public final float meanLongitude;  // Mean longitude (AngleUtils.RADIANS)
+    public final double distance;       // Mean distance (AU)
+    public final double eccentricity;   // Eccentricity of orbit
+    public final double inclination;    // Inclination of orbit (AngleUtils.RADIANS)
+    public final double ascendingNode;  // Longitude of ascending node (AngleUtils.RADIANS)
+    public final double perihelion;     // Longitude of perihelion (AngleUtils.RADIANS)
+    public final double meanLongitude;  // Mean longitude (AngleUtils.RADIANS)
 
-    public OrbitalElements(float d, float e, float i, float a, float p, float l) {
+    public OrbitalElements(double d, double e, double i, double a, double p, double l) {
         this.distance = d;
         this.eccentricity = e;
         this.inclination = i;
@@ -66,16 +65,16 @@ public class OrbitalElements {
     // m - mean anomaly in radians
     // e - orbit eccentricity
     // Return value is in radians.
-    private static float calculateTrueAnomaly(float m, float e) {
+    private static double calculateTrueAnomaly(double m, double e) {
         // initial approximation of eccentric anomaly
-        float e0 = m + e * (float) Math.sin(m) * (1.0f + e * (float) Math.cos(m));
-        float e1;
+        double e0 = m + e * (double) Math.sin(m) * (1.0f + e * (double) Math.cos(m));
+        double e1;
 
         // iterate to improve accuracy
         int counter = 0;
         do {
             e1 = e0;
-            e0 = e1 - (e1 - e * (float) Math.sin(e1) - m) / (1.0f - e * (float) Math.cos(e1));
+            e0 = e1 - (e1 - e * (double) Math.sin(e1) - m) / (1.0f - e * (double) Math.cos(e1));
             if (counter++ > 100) {
                 Log.d(TAG, "Failed to converge! Exiting.");
                 Log.d(TAG, "e1 = " + e1 + ", e0 = " + e0);
@@ -85,11 +84,11 @@ public class OrbitalElements {
         } while (abs(e0 - e1) > EPSILON);
 
         // convert eccentric anomaly to true anomaly
-        float v = 2f * (float) Math.atan((float) (Math.sqrt((1 + e) / (1 - e)) * ((float) Math.sin(0.5f * e0) / (float) Math.cos(0.5f * e0))));
+        double v = 2f * (double) Math.atan((double) (Math.sqrt((1 + e) / (1 - e)) * ((double) Math.sin(0.5f * e0) / (double) Math.cos(0.5f * e0))));
         return MathsUtils.mod2pi(v);
     }
 
-    public float getAnomaly() {
+    public double getAnomaly() {
         return calculateTrueAnomaly(meanLongitude - perihelion, eccentricity);
     }
